requirements.txt
psutil
TgCrypto
psycopg2-binary
python-telegram-bot==13.15
requests
speedtest-cli
telegraph
instaloader
pytz
get-video-properties
ffmpeg-python
pathlib2
schedule
shortuuid
python-dotenv
google-api-python-client
google-auth-oauthlib